Login update

+++ src/Components/AuthArea/Logout/Logout.tsx
@@ -7,18 +7,25 @@ import { useNavigate } from "react-router-dom";
 import store from "../../../Redux/Store";
 
 import { removeCompanies, removeCustomers } from './../../../Redux/AdminAppState';
-import { removeCoupons } from "../../../Redux/CompanyAppState";
+import { removeCompanyCoupons } from "../../../Redux/CompanyAppState";
 import { removeCustomerCoupons } from "../../../Redux/CustomerAppState";
 
 
 function Logout(): JSX.Element {
   const navigate = useNavigate();
-  useEffect(() => {
-    store.dispatch(loggedOut());
-    store.dispatch(removeCompanies());
-    store.dispatch(removeCustomers());
-    store.dispatch(removeCoupons());
-    store.dispatch(removeCustomerCoupons());
+  const userType = store.getState().userReducer.user.clientType;
+    useEffect(() => {
+        store.dispatch(loggedOut());
+        if (userType == "ADMINISTRATOR") {
+            store.dispatch(removeCompanies());
+            store.dispatch(removeCustomers());
+        }
+        if (userType == "COMPANY") {
+            store.dispatch(removeCompanyCoupons());
+        }
+        if (userType == "CUSTOMER") {
+            store.dispatch(removeCustomerCoupons());
+        }
     navigate("/home");
   }, []);
   return <></>;
